requirements.txt
numpy>=1.24.0
pyaudio>=0.2.13
librosa>=0.10.0
openai-whisper>=20231005
transformers>=4.36.0
torch>=2.1.0
pyAudioAnalysis>=0.3.14
imbalanced-learn>=0.10.0
eyed3>=0.9.6
pydub>=0.25.1
matplotlib>=3.7.0
plotly>=5.17.0
flask>=3.0.0
flask-socketio>=5.3.6
google-generativeai>=0.3.0
elevenlabs>=0.2.26
wave>=0.0.2
python-dotenv>=1.0.1
scipy>=1.11.0
soundfile>=0.12.1
webrtcvad>=2.0.10
